refactor: clean architecture - remove code duplication from integration layer
- Consolidated all types (UniversalRequest, UniversalResponse, IntegrationError) into streaming_adapter.rs
- Removed duplicate framework_helpers.rs module
- Simplified mod.rs by removing redundant exports
- Fixed imports in universal_adapter.rs to use consolidated types
- Removed TODO comments and improved framework_name() method
- All 201 tests pass after refactoring

BREAKING: Removed framework_helpers module - all functionality moved to streaming_adapter
